+++ src/Models/ItemCreators/TruffleServiceItemCreator.ts
@@ -1,10 +1,10 @@
 // Copyright (c) Consensys Software Inc. All rights reserved.
 // Licensed under the MIT license.
 
-import { AzureBlockchainService } from "../TreeItems";
-import { ItemCreator } from "./ItemCreator";
+import {AzureBlockchainService} from "../TreeItems";
+import {ItemCreator} from "./ItemCreator";
 
-export class TrufflesuiteServiceItemCreator extends ItemCreator {
+export class TruffleServiceItemCreator extends ItemCreator {
   protected createFromObject(): AzureBlockchainService {
     return new AzureBlockchainService();
   }